[project] wf : create tests/details/utils/ for tests utils
